requirements.txt
Django
django-crispy-forms
django-tinymce
fontawesomefree
django-recaptcha